+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+
+### Added
+
+- Extend association history config to support translations.
+- Support boolean values in association history.
+
 ## [5.14.1] - 2026-05-19
 
 ### Fixed
@@ -1548,6 +1555,7 @@ _Stable release based on [0.1.0-rc.1]._
 
 ## [0.0.0] - 2024-09-12
 
+[Unreleased]: https://https://github.com/internetguru/laravel-common/compare/staging...dev
 [5.14.1]: https://https://github.com/internetguru/laravel-common/compare/v5.14.0...v5.14.1
 [5.14.0]: https://https://github.com/internetguru/laravel-common/compare/v5.13.7...v5.14.0
 [5.14.0-rc.1]: https://github.com/internetguru/laravel-common/releases/tag/v5.13.7